requirements.txt
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Faker==4.0.3
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
mongoengine==0.19.1
pymongo==3.10.1
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
telebot==0.0.3
text-unidecode==1.3
timeloop==1.0.2
urllib3==1.25.8
Werkzeug==1.0.1
